api_service/requirements.txt
fastapi>=0.100.0
uvicorn==0.27.1
pydantic==1.10.13
python-dotenv>=0.19.0
aiohttp==3.9.3
slowapi==0.1.9
tenacity>=8.0.0
openai>=1.0.0
ollama>=0.1.6
fake-useragent>=1.4.0
beautifulsoup4>=4.11.1
lxml>=5.1.0
requests>=2.28.1
filelock>=3.12.0 